Misc cleanup
