+++ hedera-node/src/main/java/com/hedera/services/legacy/logic/ApplicationConstants.java
@@ -23,11 +23,9 @@ package com.hedera.services.legacy.logic;
 public class ApplicationConstants {
 	public static String NETTY_MODE_DEV = "DEV";
 
-	public static String LEDGER_PATH = "/{0}/";
 	// Default values
 	public static long DEFAULT_FEE = 100000l;
 
-	// Signature Algorithm Type Constants
 	public static String EXPORTED_ACCOUNT_PATH = "exportedAccountPath";
 	public static long EXCHANGE_RATE_FILE_ACCOUNT_NUM = 112;
 	public static long DEFAULT_FILE_SHARD = 0;
@@ -50,8 +48,6 @@ public class ApplicationConstants {
 	public static int TX_MIN_DUR = 5;
 	public static int TX_MAX_DUR = 180;
 	public static int TXN_MIN_VALIDITY_BUFFER_SECS = 10;
-
-	public static long INITIAL_GENESIS_COINS = 5000000000000000000l;
 
 	public static long DEFAULT_CONTRACT_DURATION_IN_SEC = 7890000;
